Atualizações na avaliação diagnóstica

+++ 4P/LPOO/ExerciciosDeAula/src/avaliacaoDiagnostica/VetVendedor.java
@@ -79,9 +79,9 @@ public class VetVendedor {
 	}
 	
 	public BigDecimal mediaVendas(Categoria categoria) {
-		BigDecimal quantidade = BigDecimal.valueOf(quantVendedor(categoria)).compareTo(BigDecimal.valueOf(0));
+		int quantidade = BigDecimal.valueOf(quantVendedor(categoria)).compareTo(BigDecimal.valueOf(0));
 		if (quantidade > 0)
-			return totalVendas(categoria).divide(quantidade);
+			return totalVendas(categoria).divide(BigDecimal.valueOf(quantidade));
 		return BigDecimal.valueOf(0);
 	}
